+++ NEWS.md
@@ -1,3 +1,10 @@
+# lasR 0.8.0
+
+- New: stage `sort_points()` to optimize compression.
+- New: protection against overwriting processed files while processing
+- Fix: messages and warnings are thread safe. A pipeline that printed message on the console could crash. This is expected to be safe now.
+- Fix: #70
+
 # lasR 0.7.3
 
 - Fix: #71 `write_lax()` is 2x faster when used as a single stage